fix on data type--bad data array and biplane bug--same as David's fix

+++ PYME/localization/remFitBuf.py
@@ -142,7 +142,7 @@ class CameraInfoManager(object):
         except KeyError: 
             #cache miss
             x0, y0, x1, y1 = ROI
-            mp = self._fetchMap(md, mapName)[x0:x1, y0:y1]
+            mp = self._fetchMap(md, mapName)[x0:x1, y0:y1].astype('double')
             
             self._cache[mapKey] = mp
 
@@ -569,7 +569,7 @@ class fitTask(taskDef.Task):
 
     @classmethod
     def calcSigma(cls, md, data):
-        var = np.atleast_3d(cameraMaps.getVarianceMap(md))
+        var = np.atleast_3d(cameraMaps.getVarianceMap(md)) # this must be double type!! Should we enforce with an 'astype' call?
         return np.sqrt(var + (float(md.Camera.NoiseFactor)**2)*(float(md.Camera.ElectronsPerCount)*float(md.Camera.TrueEMGain)*np.maximum(data, 1.0) + float(md.Camera.TrueEMGain)*float(md.Camera.TrueEMGain)))/float(md.Camera.ElectronsPerCount)
     
     def calcThreshold(self):
